+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Hooked up "Event Registration from Event2 detail admin settings [#3300](https://github.com/rokwire/illinois-app/issues/3300).
+
 ## [5.0.29] - 2023-07-17
 ### Changed
 - Transfer only affected profile prefs when login from onboarding [#3416](https://github.com/rokwire/illinois-app/issues/3416).